pyproject.toml
[project]
name = "langchain-chatbot-tools"
version = "0.1.0"
description = "A chatbot that uses LangChain with tools"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.1.0",
    "langchain-core>=0.1.15",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "langchainhub>=0.1.14",
    "langchain-google-genai>=0.0.2",
    "sqlite",
]
